+++ src/cpu.cpp
@@ -18,7 +18,7 @@ const std::string& CPU::modelName() const { return _modelName; }
 const std::string& CPU::vendor() const { return _vendor; }
 
 // _____________________________________________________________________________________________________________________
-int CPU::cacheSize_Bytes() const { return _cacheSize_Bytes; }
+int64_t CPU::cacheSize_Bytes() const { return _cacheSize_Bytes; }
 
 // _____________________________________________________________________________________________________________________
 int CPU::numPhysicalCores() const { return _numPhysicalCores; }
@@ -27,13 +27,13 @@ int CPU::numPhysicalCores() const { return _numPhysicalCores; }
 int CPU::numLogicalCores() const { return _numLogicalCores; }
 
 // _____________________________________________________________________________________________________________________
-int CPU::maxClockSpeed_MHz() const { return _maxClockSpeed_MHz; }
+int64_t CPU::maxClockSpeed_MHz() const { return _maxClockSpeed_MHz; }
 
 // _____________________________________________________________________________________________________________________
-int CPU::regularClockSpeed_MHz() const { return _regularClockSpeed_MHz; }
+int64_t CPU::regularClockSpeed_MHz() const { return _regularClockSpeed_MHz; }
 
 // _____________________________________________________________________________________________________________________
-int CPU::minClockSpeed_MHz() const { return _minClockSpeed_MHz; }
+int64_t CPU::minClockSpeed_MHz() const { return _minClockSpeed_MHz; }
 
 // _____________________________________________________________________________________________________________________
 const std::vector<std::string>& CPU::flags() const { return _flags; }
